move testGasLimit to test file
