+++ validation/gas_limit.go
@@ -3,12 +3,10 @@ package validation
 import (
 	"context"
 	"fmt"
-	"testing"
 
 	"github.com/ethereum-optimism/superchain-registry/superchain"
 	"github.com/ethereum-optimism/superchain-registry/validation/internal/bindings"
 	"github.com/ethereum-optimism/superchain-registry/validation/standard"
-	"github.com/stretchr/testify/require"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
@@ -16,17 +14,6 @@ import (
 )
 
 var getGasLimitWithRetriesFunc = getGasLimitWithRetries // Default implementation
-
-func testGasLimit(t *testing.T, chain *superchain.ChainConfig) {
-	rpcEndpoint := superchain.Superchains[chain.Superchain].Config.L1.PublicRPC
-	require.NotEmpty(t, rpcEndpoint)
-
-	client, err := ethclient.Dial(rpcEndpoint)
-	require.NoErrorf(t, err, "could not dial rpc endpoint %s", rpcEndpoint)
-
-	err = CheckGasLimit(chain, client)
-	require.NoError(t, err)
-}
 
 func CheckGasLimit(chain *superchain.ChainConfig, l1Client *ethclient.Client) error {
 	contractAddress, err := superchain.Addresses[chain.ChainID].AddressFor("SystemConfigProxy")
